flows/requirements.txt
pandas==1.5.2
prefect-gcp[cloud_storage]==0.2.4
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
s3fs==2023.1.0
psycopg2-binary==2.9.5
google-cloud-bigquery==3.5.0
confluent_kafka==2.0.2
pytz==2022.7.1
toml==0.10.2
kafka-python==2.0.2
python-schema-registry-client==2.4.1